+++ install/RESOURCES/APPS/Launcher/MyDelegate.h
@@ -26,5 +26,6 @@
 }
 
 - (void) openAppFolder: (id) sender;
+- (void) showInfoPanel: (id)sender;
 
 @end